fix capturing of logs for NextCloud, PHP and nginx
Normally in a Docker container we aim to output log messages to stdout or
stderr. However, for some reason this simple approach didn't seem to work for
this image, and so instead we use FIFO files for the applications to write to,
and use additional processes (managed by s6) to stream the FIFO content to the
required output stream.