create NextCloud user with given UID and GID
We specify the UID and GID of the user to run NextCloud with as environment
variables but we didn't ensure the user or group actually existed. This change
adds a check for this, and creates the 'nextcloud' user and group if they don't
already exist.

It's necessary to do this before any service processes have been started, since
they run as the named UID. I therefore had to add arkivum-run.sh, which
replaces the run.sh from the upstream Docker image. Once the user has been
checked and possibly created then the arkivum-run.sh delegates to the base
run.sh script for the main execution.

Also replaced various "COPY rootfs..." commands with single "COPY rootfs /"
command to simplify the Dockerfile slightly.