use config template MD5 to determine if config needs updating
The run script has been updated to check and record the MD5 checksum of the
config template in order to determine if the config file needs updating.
Previously we were using the existence of the config file, but this doesn't
allow for changes during upgrades whereas if an upgrade changes the
template then the config will now be updated too.

The template processing that was previously done as part of the one-time
setup is now done on every run, which is why the code has been moved into
the new arkivum-config.sh script.