check if environment has changed and trigger config update if it has
We already check if the config template has changed as a trigger for
updating the config, but we also need to check the environment variables
too, as these may impact the config.

We therefore now take a MD5 checksum of the output of `env` and store this
to a file. Then, if the env output has changed on the next run then we will
trigger a config update.