change image to use php:7.1-fpm-stretch as base instead of debian
The new base image is an official Docker Library image, so in theory should
be guaranteed to work with docker even if something changes (so will be
easier to maintain).

Part of the reason for doing this was to fix the logging, which is all now
sent to stdout and stderr, instead of being stored in log files.