don't overwrite host and protocol if env vars aren't set
We used to set some "sensible defaults" for the overwrite parameters in our
NextCloud config template. However, these were rarely right and are only
necessary if NextCloud is being hosted behind an SSL proxy, in which case
the correct values will need to be specified explicitly anyway.

So we now don't set any defaults for the overwrite parameters, and only set
them if the NEXTCLOUD_HOST and/or NEXTCLOUD_PORT env vars are given.