WIP - only append port if not default for the protocol
*** WORK IN PROGRESS - this commit is neither finished or tested! ***

We now only append the port number to the overwritehost config setting if
the port number is not the default for the protocol, i.e. not 80 for http
or not 443 for https.

We do this because otherwise the URLs NextCloud uses are different to those
that may be in the request, which breaks Shibboleth functionality.